* Movement Speed modifiers stack multiplicatively instead of additively. * Earth Shield procs get +healing bonus and modifiers from the caster instead of the target.
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@285 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/ascent-world/Unit.h
@@ -1031,7 +1031,7 @@ public:
 	uint64 m_detectRangeGUID[5];
 	int32  m_detectRangeMOD[5];
 	// Affect Speed
-	int32 m_speedModifier;
+	float m_speedModifier;
 	int32 m_slowdown;
 	float m_maxSpeed;
 	map< uint32, int32 > speedReductionMap;
